requirements.txt
openai
python-docx
python-dotenv
pypandoc
PyPDF2
fpdf
pdfplumber